move queries to context for better testing
