+++ lib/prediction_analyzer_web/templates/missed_predictions/index.html.eex
@@ -24,18 +24,18 @@
   <% end %>
   <div class=row>
     <%= cond do %>
-    <% assigns[:missing_departures_by_route_stop] -> %>
+    <% assigns[:missing_departures_for_route_stop] -> %>
       <%= link("Back to overview", to: Routes.missed_predictions_path(@conn, :index, @query_params |> Map.delete("missing_route") |> Map.delete("stop_id"))) %>
-      <%= render("missing_departures_by_route_stop.html", assigns) %>
-    <% assigns[:missing_departures_by_route] -> %>
+      <%= render("missing_departures_for_route_stop.html", assigns) %>
+    <% assigns[:missing_departures_for_route] -> %>
       <%= link("Back to overview", to: Routes.missed_predictions_path(@conn, :index, Map.delete(@query_params, "missing_route"))) %>
-      <%= render("missing_departures_by_route.html", assigns) %>
-    <% assigns[:unrealized_departures_by_route_stop] -> %>
+      <%= render("missing_departures_for_route.html", assigns) %>
+    <% assigns[:unrealized_departures_for_route_stop] -> %>
       <%= link("Back to overview", to: Routes.missed_predictions_path(@conn, :index, @query_params |> Map.delete("missed_route") |> Map.delete("stop_id"))) %>
-      <%= render("unrealized_departures_by_route_stop.html", assigns) %>
-    <% assigns[:unrealized_departures_by_route] -> %>
+      <%= render("unrealized_departures_for_route_stop.html", assigns) %>
+    <% assigns[:unrealized_departures_for_route] -> %>
       <%= link("Back to overview", to: Routes.missed_predictions_path(@conn, :index, Map.delete(@query_params, "missed_route"))) %>
-      <%= render("unrealized_departures_by_route.html", assigns) %>
+      <%= render("unrealized_departures_for_route.html", assigns) %>
     <% true -> %>
       <%= render("summary_tables.html", assigns) %>
     <% end %>
