+++ lib/prediction_analyzer_web/templates/missed_predictions/missing_departures_for_route_stop.html.eex
@@ -10,7 +10,7 @@
         </tr>
     </thead>
     <tbody>
-        <%= for {vehicle_id, trip_id, departure_time} <- @missing_departures_by_route_stop do %>
+        <%= for {vehicle_id, trip_id, departure_time} <- @missing_departures_for_route_stop do %>
         <tr>
             <td><%= vehicle_id %></td>
             <td><%= trip_id %></td>
